requirements.txt
aws-cdk-lib>=2.234.1,<3.0.0
constructs>=10.0.0,<11.0.0
PyYAML==6.0.1
typeguard==2.13.3